+++ examples/lib/commons/square_component.dart
@@ -6,9 +6,7 @@ import 'package:flame/palette.dart';
 class SquareComponent extends PositionComponent {
   Paint paint = BasicPalette.white.paint();
 
-  SquareComponent() {
-    size = Vector2.all(100.0);
-  }
+  SquareComponent() : super(size: Vector2.all(100.0));
 
   @override
   void render(Canvas c) {